# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Funnel Experts Tours UAE - API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface CreateLeadBody {
  /**
   * @minLength 2
   * @maxLength 100
   */
  name: string;
  email: string;
  /**
   * @minLength 7
   * @maxLength 20
   */
  phone: string;
  packageId?: number | null;
  packageName?: string | null;
  destination?: string | null;
  travelDate?: string | null;
  /** @minimum 1 */
  groupSize?: number | null;
  /** @maxLength 1000 */
  message?: string | null;
  source?: string | null;
}

export type UpdateLeadStatusBodyStatus =
  (typeof UpdateLeadStatusBodyStatus)[keyof typeof UpdateLeadStatusBodyStatus];

export const UpdateLeadStatusBodyStatus = {
  new: "new",
  contacted: "contacted",
  converted: "converted",
  closed: "closed",
} as const;

export interface UpdateLeadStatusBody {
  status: UpdateLeadStatusBodyStatus;
  notes?: string | null;
}

export type LeadStatus = (typeof LeadStatus)[keyof typeof LeadStatus];

export const LeadStatus = {
  new: "new",
  contacted: "contacted",
  converted: "converted",
  closed: "closed",
} as const;

export interface Lead {
  id: number;
  name: string;
  email: string;
  phone: string;
  packageId?: number | null;
  packageName?: string | null;
  destination?: string | null;
  travelDate?: string | null;
  groupSize?: number | null;
  message?: string | null;
  source?: string | null;
  status: LeadStatus;
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface LeadsListResponse {
  data: Lead[];
  total: number;
  limit: number;
  offset: number;
}

export interface Package {
  id: number;
  slug: string;
  name: string;
  tagline: string;
  description: string;
  destination: string;
  category: string;
  duration: string;
  price: number;
  currency: string;
  maxGroupSize: number;
  highlights: string[];
  includes: string[];
  excludes: string[];
  itinerary: string[];
  imageUrl?: string | null;
  badge?: string | null;
  isPopular: boolean;
  isFeatured: boolean;
  rating: number;
  reviewCount: number;
  seoTitle: string;
  seoDescription: string;
  seoKeywords: string;
  createdAt: string;
}

export interface PackagesListResponse {
  data: Package[];
  total: number;
}

export type ListLeadsParams = {
  status?: ListLeadsStatus;
  limit?: number;
  offset?: number;
};

export type ListLeadsStatus =
  (typeof ListLeadsStatus)[keyof typeof ListLeadsStatus];

export const ListLeadsStatus = {
  new: "new",
  contacted: "contacted",
  converted: "converted",
  closed: "closed",
} as const;

export type ListPackagesParams = {
  destination?: ListPackagesDestination;
  category?: ListPackagesCategory;
};

export type ListPackagesDestination =
  (typeof ListPackagesDestination)[keyof typeof ListPackagesDestination];

export const ListPackagesDestination = {
  dubai: "dubai",
  fujairah: "fujairah",
  rak: "rak",
  all: "all",
} as const;

export type ListPackagesCategory =
  (typeof ListPackagesCategory)[keyof typeof ListPackagesCategory];

export const ListPackagesCategory = {
  desert: "desert",
  city: "city",
  adventure: "adventure",
  cultural: "cultural",
  nightlife: "nightlife",
  all: "all",
} as const;