# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Funnel Experts Tours UAE - API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Submit a booking inquiry / lead
 */
export const createLeadBodyNameMin = 2;
export const createLeadBodyNameMax = 100;

export const createLeadBodyPhoneMin = 7;
export const createLeadBodyPhoneMax = 20;

export const createLeadBodyMessageMax = 1000;

export const CreateLeadBody = zod.object({
  name: zod.string().min(createLeadBodyNameMin).max(createLeadBodyNameMax),
  email: zod.string().email(),
  phone: zod.string().min(createLeadBodyPhoneMin).max(createLeadBodyPhoneMax),
  packageId: zod.number().nullish(),
  packageName: zod.string().nullish(),
  destination: zod.string().nullish(),
  travelDate: zod.string().nullish(),
  groupSize: zod.number().min(1).nullish(),
  message: zod.string().max(createLeadBodyMessageMax).nullish(),
  source: zod.string().nullish(),
});

/**
 * @summary List all leads (admin)
 */
export const ListLeadsQueryParams = zod.object({
  status: zod.enum(["new", "contacted", "converted", "closed"]).optional(),
  limit: zod.coerce.number().optional(),
  offset: zod.coerce.number().optional(),
});

export const ListLeadsResponse = zod.object({
  data: zod.array(
    zod.object({
      id: zod.number(),
      name: zod.string(),
      email: zod.string(),
      phone: zod.string(),
      packageId: zod.number().nullish(),
      packageName: zod.string().nullish(),
      destination: zod.string().nullish(),
      travelDate: zod.string().nullish(),
      groupSize: zod.number().nullish(),
      message: zod.string().nullish(),
      source: zod.string().nullish(),
      status: zod.enum(["new", "contacted", "converted", "closed"]),
      notes: zod.string().nullish(),
      createdAt: zod.date(),
      updatedAt: zod.date(),
    }),
  ),
  total: zod.number(),
  limit: zod.number(),
  offset: zod.number(),
});

/**
 * @summary Get a single lead
 */
export const GetLeadParams = zod.object({
  id: zod.coerce.number(),
});

export const GetLeadResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  phone: zod.string(),
  packageId: zod.number().nullish(),
  packageName: zod.string().nullish(),
  destination: zod.string().nullish(),
  travelDate: zod.string().nullish(),
  groupSize: zod.number().nullish(),
  message: zod.string().nullish(),
  source: zod.string().nullish(),
  status: zod.enum(["new", "contacted", "converted", "closed"]),
  notes: zod.string().nullish(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Update lead status
 */
export const UpdateLeadStatusParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateLeadStatusBody = zod.object({
  status: zod.enum(["new", "contacted", "converted", "closed"]),
  notes: zod.string().nullish(),
});

export const UpdateLeadStatusResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  phone: zod.string(),
  packageId: zod.number().nullish(),
  packageName: zod.string().nullish(),
  destination: zod.string().nullish(),
  travelDate: zod.string().nullish(),
  groupSize: zod.number().nullish(),
  message: zod.string().nullish(),
  source: zod.string().nullish(),
  status: zod.enum(["new", "contacted", "converted", "closed"]),
  notes: zod.string().nullish(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary List all tour packages
 */
export const ListPackagesQueryParams = zod.object({
  destination: zod.enum(["dubai", "fujairah", "rak", "all"]).optional(),
  category: zod
    .enum(["desert", "city", "adventure", "cultural", "nightlife", "all"])
    .optional(),
});

export const ListPackagesResponse = zod.object({
  data: zod.array(
    zod.object({
      id: zod.number(),
      slug: zod.string(),
      name: zod.string(),
      tagline: zod.string(),
      description: zod.string(),
      destination: zod.string(),
      category: zod.string(),
      duration: zod.string(),
      price: zod.number(),
      currency: zod.string(),
      maxGroupSize: zod.number(),
      highlights: zod.array(zod.string()),
      includes: zod.array(zod.string()),
      excludes: zod.array(zod.string()),
      itinerary: zod.array(zod.string()),
      imageUrl: zod.string().nullish(),
      badge: zod.string().nullish(),
      isPopular: zod.boolean(),
      isFeatured: zod.boolean(),
      rating: zod.number(),
      reviewCount: zod.number(),
      seoTitle: zod.string(),
      seoDescription: zod.string(),
      seoKeywords: zod.string(),
      createdAt: zod.date(),
    }),
  ),
  total: zod.number(),
});

/**
 * @summary Get a single package by slug
 */
export const GetPackageParams = zod.object({
  slug: zod.coerce.string(),
});

export const GetPackageResponse = zod.object({
  id: zod.number(),
  slug: zod.string(),
  name: zod.string(),
  tagline: zod.string(),
  description: zod.string(),
  destination: zod.string(),
  category: zod.string(),
  duration: zod.string(),
  price: zod.number(),
  currency: zod.string(),
  maxGroupSize: zod.number(),
  highlights: zod.array(zod.string()),
  includes: zod.array(zod.string()),
  excludes: zod.array(zod.string()),
  itinerary: zod.array(zod.string()),
  imageUrl: zod.string().nullish(),
  badge: zod.string().nullish(),
  isPopular: zod.boolean(),
  isFeatured: zod.boolean(),
  rating: zod.number(),
  reviewCount: zod.number(),
  seoTitle: zod.string(),
  seoDescription: zod.string(),
  seoKeywords: zod.string(),
  createdAt: zod.date(),
});