# lib/api-zod/src/generated/types/createLeadBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Funnel Experts Tours UAE - API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateLeadBody {
  /**
   * @minLength 2
   * @maxLength 100
   */
  name: string;
  email: string;
  /**
   * @minLength 7
   * @maxLength 20
   */
  phone: string;
  packageId?: number | null;
  packageName?: string | null;
  destination?: string | null;
  travelDate?: string | null;
  /** @minimum 1 */
  groupSize?: number | null;
  /** @maxLength 1000 */
  message?: string | null;
  source?: string | null;
}