# lib/api-zod/src/generated/types/lead.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Funnel Experts Tours UAE - API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LeadStatus } from "./leadStatus";

export interface Lead {
  id: number;
  name: string;
  email: string;
  phone: string;
  packageId?: number | null;
  packageName?: string | null;
  destination?: string | null;
  travelDate?: string | null;
  groupSize?: number | null;
  message?: string | null;
  source?: string | null;
  status: LeadStatus;
  notes?: string | null;
  createdAt: Date;
  updatedAt: Date;
}