# lib/api-zod/src/generated/types/leadsListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Funnel Experts Tours UAE - API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Lead } from "./lead";

export interface LeadsListResponse {
  data: Lead[];
  total: number;
  limit: number;
  offset: number;
}