# lib/api-zod/src/generated/types/package.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Funnel Experts Tours UAE - API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Package {
  id: number;
  slug: string;
  name: string;
  tagline: string;
  description: string;
  destination: string;
  category: string;
  duration: string;
  price: number;
  currency: string;
  maxGroupSize: number;
  highlights: string[];
  includes: string[];
  excludes: string[];
  itinerary: string[];
  imageUrl?: string | null;
  badge?: string | null;
  isPopular: boolean;
  isFeatured: boolean;
  rating: number;
  reviewCount: number;
  seoTitle: string;
  seoDescription: string;
  seoKeywords: string;
  createdAt: Date;
}