# lib/api-zod/src/generated/types/packagesListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Funnel Experts Tours UAE - API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Package } from "./package";

export interface PackagesListResponse {
  data: Package[];
  total: number;
}